third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_attributes_nohover_pointers.html

<!doctype html>
<title>Pointer Events no-hover pointer attributes test</title>
<meta name="variant" content="?touch">
<meta name="viewport" content="width=device-width">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script type="text/javascript" src="pointerevent_support.js"></script>
<style>
  div {
    height: 80px;
    width: 80px;
  }
</style>
<body onload="run()">
  <div id="square1"></div>
  <div id="done"></div>
</body>
<script>
  'use strict';
  const pointer_type = location.search.substring(1);
  const test_pointer = pointer_type + "TestPointer";

  const logged_events = [
    "pointerover", "pointerenter", "pointerdown",
    "pointerup", "pointerout", "pointerleave"
  ];

  // TODO(mustaq@chromium.org): add pointermove coverage with touch-action:none.

  function checkPointerEventAttributes(event,
      expected_pointer_id, expected_event_type,
      expected_bounding_rect) {
    assert_equals(event.pointerType, pointer_type,
        expected_event_type + ".pointerType should match test input");

    assert_equals(event.pointerId, expected_pointer_id,
        expected_event_type + ".pointerId should match all other events");
    assert_equals(event.type, expected_event_type,
        expected_event_type + ".type should be " + expected_event_type);

    assert_equals(event.button, 0,
        expected_event_type + ".button should be 0");

    if (['pointerover', 'pointerenter', 'pointerdown'].includes(event.type)) {
      assert_equals(event.buttons, 1,
          expected_event_type + ".buttons should be 1");
    } else {
      assert_equals(event.buttons, 0,
          expected_event_type + ".buttons should be 0");
    }

    assert_true(event.clientX >= expected_bounding_rect.left &&
        event.clientX < expected_bounding_rect.right,
        expected_event_type + ".clientX should be within bounding client rect.");

    assert_true(event.clientY >= expected_bounding_rect.top &&
        event.clientY < expected_bounding_rect.bottom,
        expected_event_type + ".clientY should be within bounding client rect.");

    check_PointerEvent(event, expected_event_type);

    assert_equals(event.isPrimary, true,
        expected_event_type + ".isPrimary is true.");
  }

  async function checkEventsOnTarget(test, elem) {
    const done = document.getElementById("done");

    let logged_event_promises = [];
    for (let i = 0; i < logged_events.length; i++) {
      logged_event_promises.push(getEvent(logged_events[i], elem, test));
    }
    let done_promise = getEvent("pointerup", done, test);

    let actions = new test_driver.Actions();
    actions = actions
      .addPointer(test_pointer, pointer_type)
      .pointerMove(0, 0, {origin:elem, sourceName:test_pointer})
      .pointerDown({sourceName:test_pointer})
      .pointerUp({sourceName:test_pointer})
      .pointerMove(0, 0, {origin:done, sourceName:test_pointer})
      .pointerDown({sourceName:test_pointer})
      .pointerUp({sourceName:test_pointer});
    await actions.send();

    let done_event = await done_promise;

    let expected_pointer_id;

    // All Promises in logged_event_promises are already resolved bynow.
    for (let i = 0; i < logged_event_promises.length; i++) {
      let event = await logged_event_promises[i];
      if (expected_pointer_id === undefined) {
        expected_pointer_id = event.pointerId;
      }
      checkPointerEventAttributes(event, expected_pointer_id,
          logged_events[i], elem.getBoundingClientRect());
    }
  }

  function run() {
    promise_test(async test => {
      const target = document.getElementById("square1");
      checkEventsOnTarget(test, target);
    }, "PointerEvent attributes");
  }
</script>
